requirements.txt
truthbrush
openai
python-dotenv
pytz
requests 